Configuração para banco de dados Postgresql

+++ src/main/java/com/algaworks/multidb/repository/app/LivroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.algaworks.multidb.model.app.Livro;
 
-public interface LivroRepository extends JpaRepository<Livro, Long> {
+public interface LivroRepository extends JpaRepository<Livro, Integer> {
 
 }
